# openapi/aidash.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aidash API",
    "description": "Management API for the Aidash agent mesh.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/authorization/{tenant}": {
      "get": {
        "operationId": "authorization_snapshot",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Snapshot"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "operationId": "authorization_replace",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorizationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Snapshot"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/authorization/{tenant}/catalog": {
      "get": {
        "operationId": "authorization_catalog",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Binding"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "operationId": "authorization_set_catalog",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Binding"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/authorization/{tenant}/credentials": {
      "get": {
        "operationId": "authorization_credentials",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Credential"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "operationId": "authorization_issue_credential",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedCredential"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/authorization/{tenant}/credentials/{id}/revoke": {
      "post": {
        "operationId": "authorization_revoke_credential",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Credential"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/authorization/{tenant}/decisions": {
      "get": {
        "operationId": "authorization_decisions",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {}
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/authorization/{tenant}/evaluate": {
      "post": {
        "operationId": "authorization_evaluate",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Evaluation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Decision"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/authorization/{tenant}/revisions": {
      "get": {
        "operationId": "authorization_revisions",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {}
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/authorization/{tenant}/simulate": {
      "post": {
        "operationId": "authorization_simulate",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Evaluation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Decision"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthorizationUpdate": {
        "type": "object",
        "required": ["expected_revision", "bundle"],
        "properties": {
          "bundle": {
            "$ref": "#/components/schemas/PolicyBundle"
          },
          "expected_revision": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
...
      "Binding": {
        "type": "object",
        "required": [
          "tenant",
          "entry_id",
          "entry_version",
          "enabled",
          "revision"
        ],
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "entry_id": {
            "type": "string"
          },
          "entry_version": {
            "type": "string"
          },
          "revision": {
            "type": "integer",
            "format": "int64"
          },
          "tenant": {
            "type": "string"
          }
        }
      },
      "CatalogInput": {
        "type": "object",
        "required": ["entry", "expected_revision", "enabled"],
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "entry": {
            "$ref": "#/components/schemas/EntityRef"
          },
          "expected_revision": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
...
      "Condition": {
        "oneOf": [
          {
            "type": "object",
            "required": ["conditions", "op"],
            "properties": {
              "conditions": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Condition"
                }
              },
              "op": {
                "type": "string",
                "enum": ["all"]
              }
            }
          },
          {
            "type": "object",
            "required": ["conditions", "op"],
            "properties": {
              "conditions": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Condition"
                }
              },
              "op": {
                "type": "string",
                "enum": ["any"]
              }
            }
          },
          {
            "type": "object",
            "required": ["left", "right", "op"],
            "properties": {
              "left": {
                "$ref": "#/components/schemas/Operand"
              },
              "op": {
                "type": "string",
                "enum": ["eq"]
              },
              "right": {
                "$ref": "#/components/schemas/Operand"
              }
            }
          },
          {
            "type": "object",
            "required": ["left", "right", "op"],
            "properties": {
              "left": {
                "$ref": "#/components/schemas/Operand"
              },
              "op": {
                "type": "string",
                "enum": ["not_eq"]
              },
              "right": {
                "$ref": "#/components/schemas/Operand"
              }
            }
          },
          {
            "type": "object",
            "required": ["left", "right", "op"],
            "properties": {
              "left": {
                "$ref": "#/components/schemas/Operand"
              },
              "op": {
                "type": "string",
                "enum": ["contains"]
              },
              "right": {
                "$ref": "#/components/schemas/Operand"
              }
            }
          },
          {
            "type": "object",
            "required": ["value", "op"],
            "properties": {
              "op": {
                "type": "string",
                "enum": ["exists"]
              },
              "value": {
                "$ref": "#/components/schemas/Operand"
              }
            }
          }
        ]
      },
...
      "Credential": {
        "type": "object",
        "required": [
          "id",
          "tenant",
          "subject",
          "created_at",
          "expires_at",
          "issued_by"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "issued_by": {
            "type": "string"
          },
          "revoked_at": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "subject": {
            "type": "string"
          },
          "tenant": {
            "type": "string"
          }
        }
      },
      "CredentialInput": {
        "type": "object",
        "required": ["subject"],
        "properties": {
          "expires_in_seconds": {
            "type": "integer",
            "format": "int64"
          },
          "subject": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "Decision": {
        "type": "object",
        "required": [
          "allowed",
          "reason",
          "matched_policies",
          "effective_roles",
          "revision"
        ],
        "properties": {
          "allowed": {
            "type": "boolean"
          },
          "effective_roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          },
          "matched_policies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reason": {
            "type": "string"
          },
          "revision": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "Effect": {
        "type": "string",
        "enum": ["allow", "deny"]
      },
      "EntityRef": {
        "type": "object",
        "required": ["id", "version"],
        "properties": {
          "id": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        }
      },
...
      "Evaluation": {
        "type": "object",
        "required": ["subject", "action", "resource"],
        "properties": {
          "action": {
            "type": "string"
          },
          "environment": {},
          "resource": {
            "$ref": "#/components/schemas/Resource"
          },
          "subject": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "Group": {
        "type": "object",
        "properties": {
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          }
        },
        "additionalProperties": false
      },
...
      "IssuedCredential": {
        "type": "object",
        "description": "The bearer value is returned exactly once; only its SHA-256 digest is stored.",
        "required": ["credential", "token"],
        "properties": {
          "credential": {
            "$ref": "#/components/schemas/Credential"
          },
          "token": {
            "type": "string"
          }
        }
      },
...
      "Operand": {
        "oneOf": [
          {
            "type": "object",
            "required": ["path", "source"],
            "properties": {
              "path": {
                "type": "string"
              },
              "source": {
                "type": "string",
                "enum": ["subject"]
              }
            }
          },
          {
            "type": "object",
            "required": ["path", "source"],
            "properties": {
              "path": {
                "type": "string"
              },
              "source": {
                "type": "string",
                "enum": ["resource"]
              }
            }
          },
          {
            "type": "object",
            "required": ["path", "source"],
            "properties": {
              "path": {
                "type": "string"
              },
              "source": {
                "type": "string",
                "enum": ["environment"]
              }
            }
          },
          {
            "type": "object",
            "required": ["value", "source"],
            "properties": {
              "source": {
                "type": "string",
                "enum": ["literal"]
              },
              "value": {}
            }
          }
        ]
      },
...
      "Policy": {
        "type": "object",
        "required": ["id", "effect", "subjects", "actions", "resources"],
        "properties": {
          "actions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          },
          "condition": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Condition"
              }
            ]
          },
          "effect": {
            "$ref": "#/components/schemas/Effect"
          },
          "id": {
            "type": "string"
          },
          "resources": {
            "$ref": "#/components/schemas/ResourceSelector"
          },
          "subjects": {
            "$ref": "#/components/schemas/SubjectSelector"
          }
        },
        "additionalProperties": false
      },
      "PolicyBundle": {
        "type": "object",
        "required": ["tenant"],
        "properties": {
          "groups": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/Group"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "policies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Policy"
            }
          },
          "roles": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/Role"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "subjects": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/Subject"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "tenant": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "Resource": {
        "type": "object",
        "required": ["tenant", "kind", "id"],
        "properties": {
          "attributes": {},
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "tenant": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ResourceSelector": {
        "type": "object",
        "required": ["kinds"],
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          },
          "kinds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          }
        },
        "additionalProperties": false
      },
      "Role": {
        "type": "object",
        "properties": {
          "inherits": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          }
        },
        "additionalProperties": false
      },
...
      "Snapshot": {
        "type": "object",
        "required": ["revision", "bundle"],
        "properties": {
          "bundle": {
            "$ref": "#/components/schemas/PolicyBundle"
          },
          "revision": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "Subject": {
        "type": "object",
        "required": ["kind"],
        "properties": {
          "attributes": {},
          "delegated_by": {
            "type": ["string", "null"]
          },
          "enabled": {
            "type": "boolean"
          },
          "groups": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          },
          "kind": {
            "$ref": "#/components/schemas/SubjectKind"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          }
        },
        "additionalProperties": false
      },
      "SubjectKind": {
        "type": "string",
        "enum": ["user", "agent", "node", "service"]
      },
      "SubjectSelector": {
        "type": "object",
        "properties": {
          "any": {
            "type": "boolean"
          },
          "groups": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          },
          "ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          },
          "kinds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubjectKind"
            },
            "uniqueItems": true
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}